Changed the folder structure of the gameplay project, and reorganized the includes.
