+++ Gameplay/entities/enemies/BugState.h
@@ -1,0 +1,9 @@
+#pragma once
+
+enum class BugState
+{
+	INVALID,
+	DEAD,
+	IDLE,
+	ATTACKING
+};
