+++ Gameplay/entities/player/PlayerSoundManager.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <scripting/Script.h>
-#include "PlayerState.h"
+#include "entities/player/PlayerState.h"
 
 namespace Hachiko
 {
