+++ Gameplay/misc/DynamicCamera.h
