+++ Gameplay/ui/DebugManager.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <scripting/Script.h>
-#include "Stats.h"
-#include "PlayerState.h"
+#include "entities/Stats.h"
+#include "entities/player/PlayerState.h"
 
 namespace Hachiko
 {
